+++ src/locales/es/person/index.ts
@@ -8,6 +8,7 @@ import female_prefix from './female_prefix';
 import first_name from './first_name';
 import job_area from './job_area';
 import job_descriptor from './job_descriptor';
+import job_title_pattern from './job_title_pattern';
 import job_type from './job_type';
 import last_name from './last_name';
 import last_name_pattern from './last_name_pattern';
@@ -23,6 +24,7 @@ const person: PersonDefinition = {
   first_name,
   job_area,
   job_descriptor,
+  job_title_pattern,
   job_type,
   last_name,
   last_name_pattern,