+++ src/modules/number/index.ts
@@ -1,6 +1,6 @@
-import type { SimpleFaker } from '../..';
 import { FakerError } from '../../errors/faker-error';
-import { bindThisToMemberFunctions } from '../../internal/bind-this-to-member-functions';
+import { deprecated } from '../../internal/deprecated';
+import { SimpleModuleBase } from '../../internal/module-base';
 
 /**
  * Module to generate numbers of any kind.
@@ -16,22 +16,19 @@ import { bindThisToMemberFunctions } from '../../internal/bind-this-to-member-fu
  * - For numeric strings of a given length, use [`faker.string.numeric()`](https://fakerjs.dev/api/string.html#numeric).
  * - For credit card numbers, use [`faker.finance.creditCardNumber()`](https://fakerjs.dev/api/finance.html#creditcardnumber).
  */
-export class NumberModule {
-  constructor(private readonly faker: SimpleFaker) {
-    bindThisToMemberFunctions(this);
-  }
-
+export class NumberModule extends SimpleModuleBase {
   /**
    * Returns a single random integer between zero and the given max value or the given range.
    * The bounds are inclusive.
    *
-   * @param options Maximum value or options object. Defaults to `{}`.
+   * @param options Maximum value or options object.
    * @param options.min Lower bound for generated number. Defaults to `0`.
    * @param options.max Upper bound for generated number. Defaults to `Number.MAX_SAFE_INTEGER`.
    *
-   * @throws When options define `max < min`.
+   * @throws When `min` is greater than `max`.
+   * @throws When there are no integers between `min` and `max`.
    *
-   * @see faker.string.numeric() If you would like to generate a `string` of digits with a given length (range).
+   * @see faker.string.numeric(): For generating a `string` of digits with a given length (range).
    *
    * @example
    * faker.number.int() // 2900970162509863
@@ -89,22 +86,29 @@ export class NumberModule {
   }
 
   /**
-   * Returns a single random floating-point number for a given precision or range and precision.
-   * The lower bound is inclusive, the upper bound is exclusive, unless precision is passed.
+   * Returns a single random floating-point number.
+   * The lower bound is inclusive, the upper bound is exclusive, unless `multipleOf` is passed.
    *
-   * @param options Upper bound or options object. Defaults to `{}`.
+   * @param options Upper bound or options object.
    * @param options.min Lower bound for generated number. Defaults to `0.0`.
    * @param options.max Upper bound for generated number. Defaults to `1.0`.
    * @param options.precision Precision of the generated number, for example `0.01` will round to 2 decimal points.
    * If precision is passed, the upper bound is inclusive.
+   * @param options.multipleOf The generated number will be a multiple of this property.
+   * This property can be used to limit the result to a specific number of decimal digits.
+   * For example `0.01` will round to 2 decimal points.
+   * If multipleOf is passed, the upper bound is inclusive.
+   *
+   * @throws When `min` is greater than `max`.
+   * @throws When `precision` is negative.
    *
    * @example
    * faker.number.float() // 0.5688541042618454
    * faker.number.float(3) // 2.367973240558058
    * faker.number.float({ min: -1000000 }) //-780678.849672846
    * faker.number.float({ max: 100 }) // 17.3687307164073
-   * faker.number.float({ precision: 0.1 }) // 0.9
-   * faker.number.float({ min: 10, max: 100, precision: 0.001 }) // 35.415
+   * faker.number.float({ multipleOf: 0.25 }) // 3.75
+   * faker.number.float({ min: 10, max: 100, multipleOf: 0.001 }) // 35.415
    *
    * @since 8.0.0
    */
@@ -126,8 +130,15 @@ export class NumberModule {
           max?: number;
           /**
            * Precision of the generated number.
+           *
+           * @deprecated Use `multipleOf` instead.
            */
           precision?: number;
+          /**
+           * The generated number will be a multiple of this property.
+           * If multipleOf is passed, the upper bound is inclusive.
+           */
+          multipleOf?: number;
         } = {}
   ): number {
     if (typeof options === 'number') {
@@ -136,7 +147,17 @@ export class NumberModule {
       };
     }
 
-    const { min = 0, max = 1, precision } = options;
+    // eslint-disable-next-line deprecation/deprecation
+    const { min = 0, max = 1, precision, multipleOf = precision } = options;
+
+    if (precision !== undefined) {
+      deprecated({
+        deprecated: 'faker.number.float({ precision })',
+        proposed: 'faker.number.float({ multipleOf })',
+        since: '8.4',
+        until: '9.0',
+      });
+    }
 
     if (max === min) {
       return min;
@@ -146,12 +167,18 @@ export class NumberModule {
       throw new FakerError(`Max ${max} should be greater than min ${min}.`);
     }
 
-    if (precision !== undefined) {
-      if (precision <= 0) {
-        throw new FakerError(`Precision should be greater than 0.`);
+    if (multipleOf !== undefined) {
+      if (multipleOf <= 0) {
+        // TODO @xDivisionByZerox: Clean up in v9.0
+        throw new FakerError(`multipleOf/precision should be greater than 0.`);
       }
 
-      const factor = 1 / precision;
+      const logPrecision = Math.log10(multipleOf);
+      // Workaround to get integer values for the inverse of all multiples of the form 10^-n
+      const factor =
+        multipleOf < 1 && Number.isInteger(logPrecision)
+          ? 10 ** -logPrecision
+          : 1 / multipleOf;
       const int = this.int({
         min: min * factor,
         max: max * factor,
@@ -169,13 +196,14 @@ export class NumberModule {
    * Returns a [binary](https://en.wikipedia.org/wiki/Binary_number) number.
    * The bounds are inclusive.
    *
-   * @param options Maximum value or options object. Defaults to `{}`.
+   * @param options Maximum value or options object.
    * @param options.min Lower bound for generated number. Defaults to `0`.
    * @param options.max Upper bound for generated number. Defaults to `1`.
    *
-   * @throws When options define `max < min`.
+   * @throws When `min` is greater than `max`.
+   * @throws When there are no integers between `min` and `max`.
    *
-   * @see faker.string.binary() If you would like to generate a `binary string` with a given length (range).
+   * @see faker.string.binary(): For generating a `binary string` with a given length (range).
    *
    * @example
    * faker.number.binary() // '1'
@@ -218,13 +246,14 @@ export class NumberModule {
    * Returns an [octal](https://en.wikipedia.org/wiki/Octal) number.
    * The bounds are inclusive.
    *
-   * @param options Maximum value or options object. Defaults to `{}`.
+   * @param options Maximum value or options object.
    * @param options.min Lower bound for generated number. Defaults to `0`.
    * @param options.max Upper bound for generated number. Defaults to `7`.
    *
-   * @throws When options define `max < min`.
+   * @throws When `min` is greater than `max`.
+   * @throws When there are no integers between `min` and `max`.
    *
-   * @see faker.string.octal() If you would like to generate an `octal string` with a given length (range).
+   * @see faker.string.octal(): For generating an `octal string` with a given length (range).
    *
    * @example
    * faker.number.octal() // '5'
@@ -267,11 +296,12 @@ export class NumberModule {
    * Returns a lowercase [hexadecimal](https://en.wikipedia.org/wiki/Hexadecimal) number.
    * The bounds are inclusive.
    *
-   * @param options Maximum value or options object. Defaults to `{}`.
+   * @param options Maximum value or options object.
    * @param options.min Lower bound for generated number. Defaults to `0`.
    * @param options.max Upper bound for generated number. Defaults to `15`.
    *
-   * @throws When options define `max < min`.
+   * @throws When `min` is greater than `max`.
+   * @throws When there are no integers between `min` and `max`.
    *
    * @example
    * faker.number.hex() // 'b'
@@ -314,11 +344,11 @@ export class NumberModule {
    * Returns a [BigInt](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Data_structures#bigint_type) number.
    * The bounds are inclusive.
    *
-   * @param options Maximum value or options object. Defaults to `{}`.
+   * @param options Maximum value or options object.
    * @param options.min Lower bound for generated bigint. Defaults to `0n`.
    * @param options.max Upper bound for generated bigint. Defaults to `min + 999999999999999n`.
    *
-   * @throws When options define `max < min`.
+   * @throws When `min` is greater than `max`.
    *
    * @example
    * faker.number.bigInt() // 55422n