+++ src/modules/string/index.ts
@@ -1,6 +1,5 @@
-import type { SimpleFaker } from '../..';
 import { FakerError } from '../../errors/faker-error';
-import { bindThisToMemberFunctions } from '../../internal/bind-this-to-member-functions';
+import { SimpleModuleBase } from '../../internal/module-base';
 import type { LiteralUnion } from '../../utils/types';
 
 export type Casing = 'upper' | 'lower' | 'mixed';
@@ -80,8 +79,6 @@ export type NumericChar =
 export type AlphaChar = LowerAlphaChar | UpperAlphaChar;
 export type AlphaNumericChar = AlphaChar | NumericChar;
 
-const SAMPLE_MAX_LENGTH = 2 ** 20;
-
 /**
  * Module to generate string related entries.
  *
@@ -98,11 +95,7 @@ const SAMPLE_MAX_LENGTH = 2 ** 20;
  * - Emoji can be found at [`faker.internet.emoji()`](https://fakerjs.dev/api/internet.html#emoji).
  * - The [`faker.helpers`](https://fakerjs.dev/api/helpers.html) module includes a number of string related methods.
  */
-export class StringModule {
-  constructor(private readonly faker: SimpleFaker) {
-    bindThisToMemberFunctions(this);
-  }
-
+export class StringModule extends SimpleModuleBase {
   /**
    * Generates a string from the given characters.
    *
@@ -345,7 +338,7 @@ export class StringModule {
    * @param options.length The number or range of characters to generate after the prefix. Defaults to `1`.
    * @param options.prefix Prefix for the generated number. Defaults to `'0b'`.
    *
-   * @see faker.number.binary() If you would like to generate a `binary number` (within a range).
+   * @see faker.number.binary(): For generating a binary number (within a range).
    *
    * @example
    * faker.string.binary() // '0b1'
@@ -358,6 +351,11 @@ export class StringModule {
    */
   binary(
     options: {
+      /**
+       * The number or range of characters to generate after the prefix.
+       *
+       * @default 1
+       */
       length?:
         | number
         | {
@@ -370,6 +368,11 @@ export class StringModule {
              */
             max: number;
           };
+      /**
+       * Prefix for the generated number.
+       *
+       * @default '0b'
+       */
       prefix?: string;
     } = {}
   ): string {
@@ -387,7 +390,7 @@ export class StringModule {
    * @param options.length The number or range of characters to generate after the prefix. Defaults to `1`.
    * @param options.prefix Prefix for the generated number. Defaults to `'0o'`.
    *
-   * @see faker.number.octal() If you would like to generate an `octal number` (within a range).
+   * @see faker.number.octal(): For generating an octal number (within a range).
    *
    * @example
    * faker.string.octal() // '0o3'
@@ -400,6 +403,11 @@ export class StringModule {
    */
   octal(
     options: {
+      /**
+       * The number or range of characters to generate after the prefix.
+       *
+       * @default 1
+       */
       length?:
         | number
         | {
@@ -412,6 +420,11 @@ export class StringModule {
              */
             max: number;
           };
+      /**
+       * Prefix for the generated number.
+       *
+       * @default '0o'
+       */
       prefix?: string;
     } = {}
   ): string {
@@ -530,7 +543,7 @@ export class StringModule {
    * @param options.allowLeadingZeros Whether leading zeros are allowed or not. Defaults to `true`.
    * @param options.exclude An array of digits which should be excluded in the generated string. Defaults to `[]`.
    *
-   * @see faker.number.int() If you would like to generate a `number` (within a range).
+   * @see faker.number.int(): For generating a number (within a range).
    *
    * @example
    * faker.string.numeric() // '2'
@@ -626,7 +639,7 @@ export class StringModule {
   /**
    * Returns a string containing UTF-16 chars between 33 and 125 (`!` to `}`).
    *
-   * @param length Length of the generated string. Max length is `2^20`. Defaults to `10`.
+   * @param length Length of the generated string. Defaults to `10`.
    * @param length.min The minimum number of characters to generate.
    * @param length.max The maximum number of characters to generate.
    *
@@ -652,9 +665,6 @@ export class StringModule {
         } = 10
   ): string {
     length = this.faker.helpers.rangeToNumber(length);
-    if (length >= SAMPLE_MAX_LENGTH) {
-      length = SAMPLE_MAX_LENGTH;
-    }
 
     const charCodeOption = {
       min: 33,
@@ -664,7 +674,7 @@ export class StringModule {
     let returnString = '';
 
     while (returnString.length < length) {
-      returnString += String.fromCharCode(
+      returnString += String.fromCodePoint(
         this.faker.number.int(charCodeOption)
       );
     }