+++ test/faker.spec.ts
@@ -2,6 +2,7 @@ import type { SpyInstance } from 'vitest';
 import { describe, expect, it, vi } from 'vitest';
 import { faker, Faker } from '../src';
 import { FakerError } from '../src/errors/faker-error';
+import { keys } from '../src/internal/keys';
 
 describe('faker', () => {
   it('should throw error if no locales passed', () => {
@@ -13,13 +14,13 @@ describe('faker', () => {
   });
 
   it('should not log anything on startup', () => {
-    const spies: SpyInstance[] = Object.keys(console)
+    const spies: SpyInstance[] = keys(console)
       .filter((key) => typeof console[key] === 'function')
       .map((methodName) =>
         vi.spyOn(console, methodName as keyof typeof console)
       );
 
-    // eslint-disable-next-line @typescript-eslint/no-var-requires
+    // eslint-disable-next-line @typescript-eslint/no-var-requires, unicorn/prefer-module -- Using import() requires types being build but the CI / TS-Check runs without them.
     require('..').faker;
 
     new Faker({ locale: { metadata: { title: '' } } });
@@ -41,7 +42,7 @@ describe('faker', () => {
   describe('rawDefinitions', () => {
     it('locale rawDefinition accessibility', () => {
       // Metadata
-      expect(faker.rawDefinitions.metadata.title).toBeDefined();
+      expect(faker.rawDefinitions.metadata?.title).toBeDefined();
       // Standard modules
       expect(faker.rawDefinitions.location?.city_name).toBeDefined();
       // Non-existing module
@@ -62,7 +63,7 @@ describe('faker', () => {
       // Non-existing module
       expect(faker.definitions.missing).toBeDefined();
       // Non-existing definition in a non-existing module
-      expect(() => faker.definitions.missing.missing).toThrow();
+      expect(() => faker.definitions.missing?.missing).toThrow();
       // Non-existing definition in an existing module
       expect(() => faker.definitions.location.missing).toThrow();
     });