+++ test/modules/finance.spec.ts
@@ -5,7 +5,7 @@ import { faker, fakerZH_CN } from '../../src';
 import { FakerError } from '../../src/errors/faker-error';
 import ibanLib from '../../src/modules/finance/iban';
 import { luhnCheck } from '../../src/modules/helpers/luhn-check';
-import { seededTests } from './../support/seededRuns';
+import { seededTests } from '../support/seeded-runs';
 import { times } from './../support/times';
 
 const NON_SEEDED_BASED_RUN = 5;
@@ -217,6 +217,12 @@ describe('finance', () => {
       });
 
       describe('maskedNumber()', () => {
+        it('should return contain parenthesis, ellipsis and have a length of 4 by default', () => {
+          const actual = faker.finance.maskedNumber();
+
+          expect(actual).toMatch(/\(\.{3}\d{4}\)/);
+        });
+
         it('should set a default length', () => {
           const expected = 4; // default account mask length
           const mask = faker.finance.maskedNumber({